Module slime_seed_finder::cubiomes_rs

This module was generated using the C2Rust tool and the C code from Cubiomes.

That code (with some minor modifications) was directly converted to unsafe Rust. As such, the resulting code uses pointers, calls malloc and free, and is sematically equivalent to the C code, with all bugs included. Unfortunately the C2Rust tool is still in its early phases, so the generated code has some flaws. The major flaw is that the code is duplicated instead of using use statements, so for example there is one Layer struct in each module, and they can only be converted from one to another using unsafe pointer transmute: l as *mut _ as *mut other_mod::Layer. Therefore, this module is mainly used for testing the equivalence between the pure Rust implementation from the biome_layers module and the Cubiomes implementation, and will be dropped once the pure Rust code achives similar functionality and performance.

Modules

biome_util

finders

generator

layers

rendermaplayers